[mangos-0.12] Implement ObjectGuid wrapper for guid value

+++ src/game/Map.cpp
@@ -1678,7 +1678,7 @@ float  GridMap::getLiquidLevel(float x, float y)
 uint8  GridMap::getTerrainType(float x, float y)
 {
     if (!m_liquid_type)
-        return m_liquidType;
+        return (uint8)m_liquidType;
 
     x = 16 * (32 - x/SIZE_OF_GRIDS);
     y = 16 * (32 - y/SIZE_OF_GRIDS);
@@ -1923,7 +1923,7 @@ uint16 Map::GetAreaFlag(float x, float y, float z, bool *isOutdoors) const
             *isOutdoors = true;
     }
     return areaflag;
- }
+}
 
 uint8 Map::GetTerrainType(float x, float y) const
 {
